Added additional struct argument to addy_axis, identified issue with homing, fixed issues with zooming
addy_axis now takes 'axes_struct' with an existing axes struct as an optional argument, and will add new axes to it.
'ax' is now a required argument for addy_axis, and all others have been made parameters.

Made it explicit that brushing is currently disabled.

Found an issue with homing: the Restore View callback doesn't work on additional axes anymore. Perhaps Matlab changed something in the backend.

Fixed issue with zoom due to incorrect use of the struct.